Make the 32-bit build of gcu_string_hash_64 exist
string.h declares `uint64_t gcu_string_hash_64(char const *, size_t)`.
On a 64-bit target string.c defines exactly that.  On a 32-bit target it
defined `size_t gcu_string_hash_64(...)`, and `size_t` is 32 bits there,
so the definition conflicts with its own declaration and the translation
unit does not compile.  This was not a hash that came out half as wide as
promised -- the library had no 32-bit build at all.

Nothing had ever tried one.  The arm sits behind `#if SIZE_MAX ==
18446744073709551615u`, and the arms you do not take are never parsed, so
neither -Werror nor check-clang nor any sanitizer could see it.  Found by
compiling every source for i686 and powerpc in the new container at
tools/xarch in the workspace: 27 of 28 built, this one did not, with
x86_64 building 28 of 28 as the control.  With this commit all 28 build
on i686, powerpc, s390x, sparc64 and x86_64.

Worth noting what else this means.  `gcu_string_murmur3_x86_128()`, whose
finalizer 6dfce64 corrected, has exactly one caller inside this library
and it is this function -- so the wrong hash was reachable only through
an arm that had never compiled.  That is why three years of green runs
say so little about it: the code was not being built, not merely not
being exercised.

The local also went from `int64_t buf[2]` to `uint64_t`, matching the
64-bit arm and the type actually being returned, and the 64-bit arm loses
a pointless round trip through `int64_t` on the way out.  No value
changes on any target: the three verification values and an 11-length
byte-level table are identical before and after on all seven targets in
the matrix.

No test here, because there is nothing on this host that can run one --
`make test` is 542 and unchanged.  The test is tools/xarch/murmur3.sh,
which now exercises this arm for real on i686 and powerpc.

+++ src/string.c
@@ -69,13 +69,19 @@ uint32_t gcu_string_hash_32(char const * str, size_t len) {
 uint64_t gcu_string_hash_64(char const * str, size_t len) {
   uint64_t buf[2];
   gcu_string_murmur3_x64_128(str, len, 0, buf);
-  return (int64_t)(buf[0] ^ buf[1]);
+  return buf[0] ^ buf[1];
 }
 
 #else
 
-size_t gcu_string_hash_64(char const * str, size_t len) {
-  int64_t buf[2];
+// The 32-bit arm.  Note what both the return type and the local were before:
+// `size_t` and `int64_t`.  `size_t` is 32 bits on a 32-bit target, and
+// string.h promises uint64_t, so this did not merely truncate the hash to
+// half its width -- it did not compile at all.  Every 32-bit build of this
+// library failed on this one function, and no build had ever been attempted:
+// see notes/cutil/murmur3-correctness.md.
+uint64_t gcu_string_hash_64(char const * str, size_t len) {
+  uint64_t buf[2];
   gcu_string_murmur3_x86_128(str, len, 0, buf);
   return buf[0] ^ buf[1];
 }
